feat feedgram: add rss reader and tg message dispatcher

+++ internal/components/storage/get_sources.go
@@ -8,7 +8,7 @@ import (
 	"github.com/IlyaZh/feedsgram/internal/queries"
 )
 
-func (c *Component) GetSource(ctx context.Context, id *int, isActive *bool, limit *int) ([]entities.Source, error) {
+func (c *Component) GetSource(ctx context.Context, id *int64, isActive *bool, limit *int) ([]entities.Source, error) {
 	settings := c.configs.GetValues().Mysql
 	if limit != nil {
 		limit = &settings.Limit
